// delcitro-app/src/componets/models/empleados.model.ts
export class UpdateEmpleadoDto {
  primer_nombre?: string;
  segundo_nombre?: string;
  apellido_paterno?: string;
  apellido_materno?: string;
  correo?: string;
  telefono?: string;
  direccion?: string;
  edad?: number;
  genero?: string;
  status?: string;
  area_id?: number;
  departamento_id?: number;
  puesto_id?: number;
}